registered add for more interfaces
