+++ torch2trt/converters/__init__.py
@@ -3,7 +3,6 @@ from .AdaptiveAvgPool2d import *
 from .add import *
 from .mul import *
 from .div import *
-from .iadd import *
 from .AvgPool2d import *
 from .BatchNorm1d import *
 from .BatchNorm2d import *
