+++ torch2trt/converters/add.py
@@ -2,6 +2,8 @@ from torch2trt.torch2trt import *
 from torch2trt.module_test import add_module_test
 
 
+@tensorrt_converter('torch.add')
+@tensorrt_converter('torch.Tensor.__iadd__')
 @tensorrt_converter('torch.Tensor.__add__')
 def convert_add(ctx):
     input_a = ctx.method_args[0]
@@ -21,3 +23,30 @@ class Add(torch.nn.Module):
 @add_module_test(torch.float32, torch.device('cuda'), [(1, 3, 224, 224), (1, 3, 224, 224)])
 def test_add_basic():
     return Add()
+
+
+class IAdd(torch.nn.Module):
+    def __init__(self):
+        super(IAdd, self).__init__()
+
+    def forward(self, x, y):
+        x += y
+        return x
+
+
+@add_module_test(torch.float32, torch.device('cuda'), [(1, 3, 224, 224), (1, 3, 224, 224)])
+def test_add_iadd():
+    return IAdd()
+
+
+class TorchAdd(torch.nn.Module):
+    def __init__(self):
+        super(TorchAdd, self).__init__()
+
+    def forward(self, x, y):
+        return torch.add(x, y)
+
+
+@add_module_test(torch.float32, torch.device('cuda'), [(1, 3, 224, 224), (1, 3, 224, 224)])
+def test_add_torchadd():
+    return TorchAdd()
