--- torch2trt/converters/iadd.py
@@ -1,24 +1,0 @@
-from torch2trt.torch2trt import *
-from torch2trt.module_test import add_module_test
-
-
-@tensorrt_converter('torch.Tensor.__iadd__')
-def convert_iadd(ctx):
-    input_a = ctx.method_args[0]
-    input_b = ctx.method_args[1]
-    layer = ctx.network.add_elementwise(input_a._trt, input_b._trt, trt.ElementWiseOperation.SUM)
-    ctx.method_args[0]._trt = layer.get_output(0)
-
-
-class IAdd(torch.nn.Module):
-    def __init__(self):
-        super(IAdd, self).__init__()
-
-    def forward(self, x, y):
-        x += y
-        return x
-
-
-@add_module_test(torch.float32, torch.device('cuda'), [(1, 3, 224, 224), (1, 3, 224, 224)])
-def test_iadd_basic():
-    return IAdd()
